// src/app/models/Cliente.ts
export class Cliente {

    id: string | undefined;
    nombre: string | undefined;
    apellidos: string| undefined;
    genero: string | undefined;
    telefono: string| undefined;
    email: string| undefined;
    contraseña: string| undefined;
    direccion: string| undefined;
    ocupacion: string| undefined;
    id_rol: number| undefined;
    created_at: string| undefined;
}